requirements.txt
python-telegram-bot==21.0
httpx>=0.27.0
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
